REFACTOR :: getTotalPrice method 추출 (#36)
- method 추출해서 order, cart에서 재사용하도록 수정
