+++ src/pages/Order/Order.scss
@@ -226,6 +226,11 @@ section {
         line-height: 29px;
         text-align: left;
       }
+
+      .order-price-area--order-price-box--box {
+        padding: 17px 16px 18px 18px;
+        background-color: #fafafa;
+      }
     }
   }
